+++ src/filters/reframe_mp3.c
@@ -197,9 +197,9 @@ static void mp3_dmx_id3_set_string(GF_MP3DmxCtx *ctx, char *name, u8 *buf)
 {
 	if ((buf[0]==0xFF) || (buf[0]==0xFE)) {
 		const u16 *sptr = (u16 *) (buf+2);
-		s32 len = UTF8_MAX_BYTES_PER_CHAR*gf_utf8_wcslen(sptr);
+		s32 len = UTF8_MAX_BYTES_PER_CHAR * (u32) gf_utf8_wcslen(sptr);
 		char *tmp = gf_malloc(len+1);
-		len = gf_utf8_wcstombs(tmp, len, &sptr);
+		len = (s32) gf_utf8_wcstombs(tmp, len, &sptr);
 		if (len>=0) {
 			tmp[len] = 0;
 			gf_filter_pid_set_property_dyn(ctx->opid, name, &PROP_STRING(tmp) );
